add metrics and move functions into files
