+++ pkg/execution/state/redis_state/queue.go
@@ -3098,251 +3098,6 @@ func (q *queue) peekGlobalNormalizeAccounts(ctx context.Context, until time.Time
 	return p.peekUUIDPointer(ctx, rc.kg.GlobalAccountNormalizeSet(), true, until, limit)
 }
 
-func (q *queue) ShadowPartitionPeek(ctx context.Context, sp *QueueShadowPartition, sequential bool, until time.Time, limit int64) ([]*QueueBacklog, int, error) {
-	if q.primaryQueueShard.Kind != string(enums.QueueShardKindRedis) {
-		return nil, 0, fmt.Errorf("unsupported queue shard kind for ShadowPartitionPeek: %s", q.primaryQueueShard.Kind)
-	}
-
-	rc := q.primaryQueueShard.RedisClient
-
-	shadowPartitionSet := rc.kg.ShadowPartitionSet(sp.PartitionID)
-
-	p := peeker[QueueBacklog]{
-		q:               q,
-		opName:          "ShadowPartitionPeek",
-		keyMetadataHash: q.primaryQueueShard.RedisClient.kg.BacklogMeta(),
-		max:             ShadowPartitionPeekMaxBacklogs,
-		maker: func() *QueueBacklog {
-			return &QueueBacklog{}
-		},
-		handleMissingItems: func(pointers []string) error {
-			err := rc.Client().Do(ctx, rc.Client().B().Zrem().Key(shadowPartitionSet).Member(pointers...).Build()).Error()
-			if err != nil {
-				q.logger.Warn().
-					Interface("missing", pointers).
-					Interface("sp", sp).
-					Msg("failed to clean up dangling backlogs from shard partition")
-			}
-
-			return nil
-		},
-	}
-
-	res, err := p.peek(ctx, shadowPartitionSet, sequential, until, limit)
-	if err != nil {
-		if errors.Is(err, ErrPeekerPeekExceedsMaxLimits) {
-			return nil, 0, ErrShadowPartitionBacklogPeekMaxExceedsLimits
-		}
-		return nil, 0, fmt.Errorf("could not peek shadow partition backlogs: %w", err)
-	}
-
-	return res.Items, res.TotalCount, nil
-}
-
-func (q *queue) ShadowPartitionPeekNormalizeBacklogs(ctx context.Context, sp *QueueShadowPartition, limit int64) ([]*QueueBacklog, error) {
-	if q.primaryQueueShard.Kind != string(enums.QueueShardKindRedis) {
-		return nil, fmt.Errorf("unsupported queue shard kind for ShadowPartitionPeekNormalizeBacklogs: %s", q.primaryQueueShard.Kind)
-	}
-
-	rc := q.primaryQueueShard.RedisClient
-
-	partitionNormalizeSet := rc.kg.PartitionNormalizeSet(sp.PartitionID)
-
-	p := peeker[QueueBacklog]{
-		q:               q,
-		opName:          "ShadowPartitionPeekNormalizeBacklogs",
-		keyMetadataHash: q.primaryQueueShard.RedisClient.kg.BacklogMeta(),
-		max:             NormalizePartitionPeekMax,
-		maker: func() *QueueBacklog {
-			return &QueueBacklog{}
-		},
-		handleMissingItems: func(pointers []string) error {
-			err := rc.Client().Do(ctx, rc.Client().B().Zrem().Key(partitionNormalizeSet).Member(pointers...).Build()).Error()
-			if err != nil {
-				q.logger.Warn().
-					Interface("missing", pointers).
-					Interface("sp", sp).
-					Msg("failed to clean up dangling backlog pointers from shadow partition normalize set")
-			}
-
-			return nil
-		},
-		// faster option: load items regardless of zscore
-		ignoreUntil: true,
-	}
-
-	res, err := p.peek(ctx, partitionNormalizeSet, false, q.clock.Now(), limit)
-	if err != nil {
-		return nil, fmt.Errorf("could not peek backlogs for normalization: %w", err)
-	}
-
-	return res.Items, nil
-}
-
-func (q *queue) BacklogNormalizePeek(ctx context.Context, b *QueueBacklog, limit int64) (*peekResult[osqueue.QueueItem], error) {
-	if q.primaryQueueShard.Kind != string(enums.QueueShardKindRedis) {
-		return nil, fmt.Errorf("unsupported queue shard kind for BacklogNormalizePeek: %s", q.primaryQueueShard.Kind)
-	}
-
-	rc := q.primaryQueueShard.RedisClient
-
-	backlogSet := rc.kg.BacklogSet(b.BacklogID)
-
-	p := peeker[osqueue.QueueItem]{
-		q:               q,
-		opName:          "BacklogNormalizePeek",
-		keyMetadataHash: q.primaryQueueShard.RedisClient.kg.QueueItem(),
-		max:             NormalizePartitionPeekMax,
-		maker: func() *osqueue.QueueItem {
-			return &osqueue.QueueItem{}
-		},
-		handleMissingItems: func(pointers []string) error {
-			err := rc.Client().Do(ctx, rc.Client().B().Zrem().Key(backlogSet).Member(pointers...).Build()).Error()
-			if err != nil {
-				q.logger.Warn().
-					Interface("missing", pointers).
-					Interface("backlog", b).
-					Msg("failed to clean up dangling queue items from backlog")
-			}
-
-			return nil
-		},
-		// faster option: load items regardless of zscore
-		ignoreUntil: true,
-	}
-
-	res, err := p.peek(ctx, backlogSet, false, q.clock.Now(), limit)
-	if err != nil {
-		return nil, fmt.Errorf("could not peek backlog items for normalization: %w", err)
-	}
-
-	return res, nil
-}
-
-func (q *queue) ShadowPartitionExtendLease(ctx context.Context, sp *QueueShadowPartition, leaseID ulid.ULID, duration time.Duration) (*ulid.ULID, error) {
-	ctx = redis_telemetry.WithScope(redis_telemetry.WithOpName(ctx, "ShadowPartitionExtendLease"), redis_telemetry.ScopeQueue)
-
-	if q.primaryQueueShard.Kind != string(enums.QueueShardKindRedis) {
-		return nil, fmt.Errorf("unsupported queue shard kind for ShadowPartitionExtendLease: %s", q.primaryQueueShard.Kind)
-	}
-
-	now := q.clock.Now()
-	leaseExpiry := now.Add(duration)
-	newLeaseID, err := ulid.New(ulid.Timestamp(leaseExpiry), rand.Reader)
-	if err != nil {
-		return nil, fmt.Errorf("could not generate new leaseID: %w", err)
-	}
-
-	sp.LeaseID = &newLeaseID
-
-	accountID := uuid.Nil
-	if sp.AccountID != nil {
-		accountID = *sp.AccountID
-	}
-
-	keys := []string{
-		q.primaryQueueShard.RedisClient.kg.ShadowPartitionMeta(),
-		q.primaryQueueShard.RedisClient.kg.GlobalShadowPartitionSet(),
-		q.primaryQueueShard.RedisClient.kg.GlobalAccountShadowPartitions(),
-		q.primaryQueueShard.RedisClient.kg.AccountShadowPartitions(accountID),
-	}
-	args, err := StrSlice([]any{
-		sp.PartitionID,
-		accountID,
-		leaseID,
-		newLeaseID,
-		now.UnixMilli(),
-		leaseExpiry.Unix(),
-	})
-	if err != nil {
-		return nil, fmt.Errorf("could not serialize args: %w", err)
-	}
-
-	status, err := scripts["queue/shadowPartitionExtendLease"].Exec(
-		redis_telemetry.WithScriptName(ctx, "shadowPartitionExtendLease"),
-		q.primaryQueueShard.RedisClient.unshardedRc,
-		keys,
-		args,
-	).AsInt64()
-	if err != nil {
-		return nil, fmt.Errorf("error extending shadow partition lease: %w", err)
-	}
-	switch status {
-	case 0:
-		return &newLeaseID, nil
-	case -1:
-		return nil, ErrShadowPartitionNotFound
-	case -2:
-		return nil, ErrShadowPartitionLeaseNotFound
-	case -3:
-		return nil, ErrShadowPartitionAlreadyLeased
-	case -4:
-		return nil, ErrShadowPartitionPaused
-	default:
-		return nil, fmt.Errorf("unknown response extending shadow partition lease: %v (%T)", status, status)
-	}
-}
-
-func (q *queue) ShadowPartitionRequeue(ctx context.Context, sp *QueueShadowPartition, leaseID ulid.ULID, requeueAt *time.Time) error {
-	ctx = redis_telemetry.WithScope(redis_telemetry.WithOpName(ctx, "ShadowPartitionRequeue"), redis_telemetry.ScopeQueue)
-
-	if q.primaryQueueShard.Kind != string(enums.QueueShardKindRedis) {
-		return fmt.Errorf("unsupported queue shard kind for ShadowPartitionRequeue: %s", q.primaryQueueShard.Kind)
-	}
-
-	sp.LeaseID = nil
-
-	accountID := uuid.Nil
-	if sp.AccountID != nil {
-		accountID = *sp.AccountID
-	}
-
-	var requeueAtS int64
-	if requeueAt != nil {
-		requeueAtS = requeueAt.Unix()
-	}
-
-	keys := []string{
-		q.primaryQueueShard.RedisClient.kg.ShadowPartitionMeta(),
-		q.primaryQueueShard.RedisClient.kg.GlobalShadowPartitionSet(),
-		q.primaryQueueShard.RedisClient.kg.GlobalAccountShadowPartitions(),
-		q.primaryQueueShard.RedisClient.kg.AccountShadowPartitions(accountID),
-		q.primaryQueueShard.RedisClient.kg.ShadowPartitionSet(sp.PartitionID),
-	}
-	args, err := StrSlice([]any{
-		sp.PartitionID,
-		accountID,
-		leaseID,
-		q.clock.Now().UnixMilli(),
-		requeueAtS,
-	})
-	if err != nil {
-		return fmt.Errorf("could not serialize args: %w", err)
-	}
-
-	status, err := scripts["queue/shadowPartitionRequeue"].Exec(
-		redis_telemetry.WithScriptName(ctx, "shadowPartitionRequeue"),
-		q.primaryQueueShard.RedisClient.unshardedRc,
-		keys,
-		args,
-	).AsInt64()
-	if err != nil {
-		return fmt.Errorf("error returning shadow partition lease: %w", err)
-	}
-	switch status {
-	case 0:
-		return nil
-	case -1:
-		return ErrShadowPartitionNotFound
-	case -2:
-		return ErrShadowPartitionAlreadyLeased
-	case -3:
-		return ErrShadowPartitionLeaseNotFound
-	default:
-		return fmt.Errorf("unknown response returning shadow partition lease: %v (%T)", status, status)
-	}
-}
-
 // PartitionLease leases a partition for a given workflow ID.  It returns the new lease ID.
 //
 // NOTE: This does not check the queue/partition name against allow or denylists;  it assumes
